ajustes en el backEnd

+++ backend/src/main/java/cr/una/consultores/dto/ResultadoControlDTO.java
@@ -5,7 +5,7 @@ public class ResultadoControlDTO {
     public String codigo;
     public String nombre;
     public String dominio;
-    public int nivelMadurez;
+    public double nivelMadurez;   // ahora es promedio (puede ser decimal, ej. 3.5)
     public double cumplimiento;
     public double riesgoControl;
     public boolean afectaC;
